Refactor callback handling: move send_callback function to utils and clean up imports
