+++ backend/app/services/response/callbacks.py
@@ -1,8 +1,4 @@
-from app.utils import APIResponse
-import requests
-import logging
-
-logger = logging.getLogger(__name__)
+from app.utils import APIResponse, send_callback
 
 
 def get_additional_data(request: dict) -> dict:
@@ -21,21 +17,6 @@ def get_additional_data(request: dict) -> dict:
     else:
         exclude_keys = sync_exclude_keys
     return {k: v for k, v in request.items() if k not in exclude_keys}
-
-
-def send_callback(callback_url: str, data: dict):
-    """Send results to the callback URL (synchronously)."""
-    try:
-        session = requests.Session()
-        # uncomment this to run locally without SSL
-        # session.verify = False
-        response = session.post(callback_url, json=data)
-        response.raise_for_status()
-        logger.info(f"[send_callback] Callback sent successfully to {callback_url}")
-        return True
-    except requests.RequestException as e:
-        logger.error(f"[send_callback] Callback failed: {str(e)}", exc_info=True)
-        return False
 
 
 def send_response_callback(
