+++ backend/app/utils.py
@@ -3,6 +3,7 @@ import logging
 from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
+import requests
 from typing import Any, Dict, Generic, Optional, TypeVar
 
 import jwt
@@ -219,6 +220,21 @@ def handle_openai_error(e: openai.OpenAIError) -> str:
         except:
             pass
     return str(e)
+
+
+def send_callback(callback_url: str, data: dict):
+    """Send results to the callback URL (synchronously)."""
+    try:
+        session = requests.Session()
+        # uncomment this to run locally without SSL
+        # session.verify = False
+        response = session.post(callback_url, json=data)
+        response.raise_for_status()
+        logger.info(f"[send_callback] Callback sent successfully to {callback_url}")
+        return True
+    except requests.RequestException as e:
+        logger.error(f"[send_callback] Callback failed: {str(e)}", exc_info=True)
+        return False
 
 
 @ft.singledispatch
